changed data type of 'value' from Long to Double

+++ core/src/main/java/com/huawei/openstack4j/openstack/cdn/v1/domain/DomainSummaryDetail.java
@@ -48,7 +48,7 @@ public class DomainSummaryDetail implements ModelEntity{
 	@JsonProperty("stat_type")
 	private String statType;
 	
-	private List<Long> values;
+	private List<Double> values;
 	
 	@JsonProperty("service_area")
 	private String serviceArea;
